pyproject.toml
[project]
name = "opsmith-cli"
version = "0.1.1"
description = "Opsmith is an AI devops engineer in your terminal"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python",
    "Topic :: Software Development",
]
license = "GPL-3.0-only"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "boto3-stubs==1.38.36",
    "gitpython>=3.1.44",
    "google-cloud-resource-manager>=1.14.2",
    "grep-ast>=0.9.0",
    "inquirer>=3.4.0",
    "networkx>=3.4.2",
    "pick>=2.4.0",
    "pydantic>=2.11.5",
    "pydantic-ai[logfire]>=0.2.16",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
    "tree-sitter-language-pack>=0.7.3",
    "typer>=0.16.0",
]

[project.scripts]
opsmith = "opsmith.main:app"

[tool.setuptools]
include-package-data = true

[tool.hatch.build.targets.wheel]
packages = ["opsmith"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
